requirements.txt
django==4.2.11
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5
django-environ==0.11.2
mysqlclient==2.2.1
PyJWT==2.8.0
python-dotenv==1.0.0 